closes #72
