+++ config/environments/default.json
@@ -1,6 +1,6 @@
 {
   "port": 3000,
-  "timeout": 60000,
+  "timeout": 6000,
   "requestIdHeader": "X-Request-Id",
   "proxy": false,
   "frontend": { "url": "localhost:3000" },
