+++ lib/mail.js
@@ -1,10 +1,9 @@
 /* eslint-disable no-undef */
 const nodemailer = require('nodemailer')
 const config = require('../config')
-const logger = require('./logger')
 
 const transport = nodemailer.createTransport(
-  Object.assign(config.get('mail:smtp'), { logger }),
+  config.get('mail:smtp'),
   config.get('mail:defaults')
 )
