+++ lib/queue.js
@@ -1,5 +1,6 @@
 const Bull = require('bull')
 const { createClient } = require('./redis')
+const merge = require('lodash/merge')
 const log = require('./logger')
 const config = require('../config')
 
@@ -9,7 +10,7 @@ const queue = new Bull(
 )
 const mailQueue = new Bull(
   'mail',
-  Object.assign(config.get('queue'), config.get('mailQueue'), { createClient })
+  merge(config.get('queue'), config.get('mailQueue'))
 )
 
 ;[queue, mailQueue].forEach(queue => {
