+++ middlewares/session.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 const session = require('express-session')
 const RedisStore = require('connect-redis')(session)
 const client = require('../lib/redis')
@@ -10,6 +11,12 @@ module.exports = session(
     store: new RedisStore({ client }),
     // this is to allow search by user id
     // see: https://github.com/jaredhanson/passport/blob/master/lib/sessionmanager.js#L21
-    genid: req => `${get(req, '_passport.session.user.id')}:${nanoid(10)}`
+    genid: req => {
+      // console.log('GENERATING SESSION ID FROM REQUEST')
+      // console.log(req._passport)
+      // console.log('==================================')
+      // return `${get(req, '_passport.session.user.id')}:${nanoid(10)}`
+      return nanoid(10)
+    }
   })
 )
