+++ package.json
@@ -16,13 +16,13 @@
     "redis",
     "api",
     "backend",
-    "json schema",
+    "jsonschema",
     "validation",
     "email",
     "template",
     "handlebars",
     "mjml",
-    "access control",
+    "access-control",
     "ACL",
     "RBAC",
     "roles",
@@ -38,7 +38,7 @@
     "start:worker": "pm2-runtime start pm2.config.js --only worker",
     "dev": "run-p dev:* watch",
     "dev:web": "nodemon server",
-    "dev:worker": "nodemon worker",
+    "dev:worker": "wait-on http://localhost:1080 && nodemon worker",
     "dev:mail": "maildev",
     "migrate": "knex migrate:latest",
     "migrate:make": "knex migrate:make",
@@ -164,7 +164,7 @@
     "nodemailer": "^6.2.1",
     "normalize-email": "^1.1.1",
     "objection": "^1.6.9",
-    "objection-authorize": "^0.3.1",
+    "objection-authorize": "^0.5.3",
     "objection-db-errors": "^1.1.1",
     "objection-password": "^2.0.0",
     "objection-table-name": "^1.0.0",
