+++ routes/index.js
@@ -4,4 +4,3 @@ module.exports = Router()
   .get('/', (req, res) => res.sendStatus(200))
   .use(require('./sessions'))
   .use(require('./users'))
-  .use(require('./account'))
