+++ views/emails/password-reset/html.hbs
@@ -94,7 +94,7 @@
                 <table border="0" cellpadding="0" cellspacing="0" role="presentation" style="vertical-align:top;" width="100%">
                   <tr>
                     <td align="left" style="font-size:0px;padding:10px 25px;word-break:break-word;">
-                      <div style="font-family:Helvetica;font-size:18px;line-height:1;text-align:left;color:#000000;">Hi {{user.username}}, We received a request to reset your password. To set a new password, click this button:</div>
+                      <div style="font-family:Helvetica;font-size:18px;line-height:1;text-align:left;color:#000000;">Hi {{user.username}}, We received a request to reset your password. <br /> To set a new password, click this button:</div>
                     </td>
                   </tr>
                   <tr>
