+++ views/emails/verify-email/html.hbs
@@ -94,7 +94,7 @@
                 <table border="0" cellpadding="0" cellspacing="0" role="presentation" style="vertical-align:top;" width="100%">
                   <tr>
                     <td align="left" style="font-size:0px;padding:10px 25px;word-break:break-word;">
-                      <div style="font-family:Helvetica;font-size:18px;line-height:1;text-align:left;color:#000000;">Hi {{user.username}}, thanks for signing up! To finish setting up your account, we just need you to verify your email address by clicking this button:</div>
+                      <div style="font-family:Helvetica;font-size:18px;line-height:1;text-align:left;color:#000000;">Hi {{user.username}}, thanks for signing up! <br /> To finish setting up your account, we just need you to verify your email address by clicking this button:</div>
                     </td>
                   </tr>
                   <tr>
